Split out commands and command switch
